server/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0.3
django-ckeditor==6.7.3
django-ckeditor-5==0.2.20
django-cloudinary-storage==0.3.0
django-environ==0.13.0
django-filter==25.2
django-js-asset==3.1.2
djangorestframework==3.17.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
